requirement.txt
python 3.11
pyperclip-1.8.2